move src/cclib -> cclib, src/scripts -> cclib/scripts
