+++ cclib/bridge/cclib2openbabel.py
