+++ cclib/bridge/cclib2pyquante.py
