+++ cclib/io/ccio.py
