+++ cclib/method/density.py
