+++ cclib/method/electrons.py
