+++ cclib/method/fragments.py
