+++ cclib/method/opa.py
