+++ cclib/parser/gamessparser.py
