+++ cclib/parser/gamessukparser.py
