+++ cclib/parser/psiparser.py
