+++ cclib/parser/utils.py
