+++ cclib/progress/qt4progress.py
