+++ cclib/progress/textprogress.py
