+++ cclib/scripts/__init__.py
