+++ cclib/scripts/ccwrite.py
